+++ packages/compilers/src/lib/common.ts
@@ -74,9 +74,15 @@ export function asyncExec(
         if (error) {
           reject(error);
         } else if (stderr) {
-          reject(
-            new Error(`Compiler process returned with errors:\n ${stderr}`),
-          );
+          // Vyper compilers <0.4.0 outputs warnings to stderr
+          // we handle this by checking if the stderr starts with "Warning:"
+          if (stderr.startsWith('Warning:')) {
+            resolve(stdout);
+          } else {
+            reject(
+              new Error(`Compiler process returned with errors:\n ${stderr}`),
+            );
+          }
         } else {
           resolve(stdout);
         }