+++ packages/lib-sourcify/src/Compilation/CompilationTypes.ts
@@ -35,6 +35,7 @@ export type CompilationLanguage = 'Solidity' | 'Vyper';
 export type CompilationErrorCode =
   | 'cannot_generate_cbor_auxdata_positions'
   | 'invalid_compiler_version'
+  | 'unsupported_compiler_version'
   | 'contract_not_found_in_compiler_output'
   | 'compiler_error'
   | 'no_compiler_output'