+++ services/server/test/unit/utils/parsing-util.spec.ts
@@ -1,29 +1,22 @@
-import { getContractPathFromSources } from "../../../src/server/services/utils/parsing-util";
 import { expect } from "chai";
+import { getContractPathFromSources } from "../../../src/server/services/utils/parsing-util";
 
-describe("getContractPathFromSources", () => {
+describe("getContractPathFromSources (lib)", () => {
   it("should return the correct contract path", () => {
     const sources = {
-      // It should not derive it from the path
-      "SolidityContract.sol": {
-        content: "contract WrongContract {}",
-      },
+      "SolidityContract.sol": { content: "contract WrongContract {}" },
       "path/file.sol": {
         content: "contract WrongContract {}\ncontract SolidityContract {}",
       },
-    };
-
+    } as any;
     const result = getContractPathFromSources("SolidityContract", sources);
     expect(result).to.equal("path/file.sol");
   });
 
   it("should return undefined when the contract path is not found in the provided sources", () => {
     const sources = {
-      "path/file.sol": {
-        content: "contract SolidityContract {}",
-      },
-    };
-
+      "path/file.sol": { content: "contract SolidityContract {}" },
+    } as any;
     const result = getContractPathFromSources(
       "AnotherSolidityContract",
       sources,