+++ services/monitor/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## sourcify-monitor@1.3.19 - 2025-08-25
+
+- Update dependencies
+
 ## sourcify-monitor@1.3.18 - 2025-08-12
 
 - Update Dockerfiles from Bullseye to Bookworm for Vyper compiler compatibility (#2289)